Google Drive bug
When you only want to export to Google drive, it also downloads it. This fixes the issue.
I also want to mention that It was too hard to get it working because there was no documentation how to contribute to the project. Maybe can you add that? I'm going to open an issue about this.